requirements.txt
mysql<2
mysqlclient<3
mysql-connector-python<9
python-dotenv<2
pandas<3
